backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg[binary]==3.1.18
python-dotenv==1.0.0
django-cors-headers==4.3.1
pyotp==2.9.0
setuptools==70.0.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

